+++ GianMarco/src/Main.cs
@@ -3,11 +3,11 @@ class Program
 {
 	static int Main(string[] args)
 	{
-		if (args.Length == 3)
+		if (args.Length == 3) // for genetic algorithm
 		{
-			Search.Utils.MoveOrdering.CaptureBonus = short.Parse(args[0]);
-			Search.Utils.MoveOrdering.PromotionBonus = short.Parse(args[1]);
-			Search.Utils.MoveOrdering.CastleBonus = short.Parse(args[2]);
+			Search.Utils.MoveOrdering.CaptureBonus = int.Parse(args[0]);
+			Search.Utils.MoveOrdering.PromotionBonus = int.Parse(args[1]);
+			Search.Utils.MoveOrdering.CastleBonus = int.Parse(args[2]);
 		}
 
 		UCI.PreBuiltInterpreter.RunAndDelegateCommands();
